[Cherry-pick] CodeQL Fixes - Second Pass from 202202
Makes changes to comply with alerts raised by CodeQL.

A prior pass with a subset of queries used here was already done so this
is mostly addressing issues discovered by enabling a few new queries.

Most of the issues here fall into the following two categories:

1. Potential use of uninitialized pointer
2. Inconsistent integer width used in loop comparison

- [ ] Breaking change?
- Will this change break pre-existing builds or functionality without
action being taken?
  **No** - Should just be refactoring to address alerts

1. Verified CI build
2. Verified boot to EFI shell on QemuQ35Pkg

N/A

(cherry picked from commit 43e657c)

+++ MdeModulePkg/Universal/PCD/Pei/Pcd.c
@@ -574,7 +574,7 @@ PeiPcdSetSku (
   }
 
   SkuIdTable = (SKU_ID *)((UINT8 *)PeiPcdDb + PeiPcdDb->SkuIdTableOffset);
-  for (Index = 0; Index < SkuIdTable[0]; Index++) {
+  for (Index = 0; (UINT64)Index < SkuIdTable[0]; Index++) {
     if (SkuId == SkuIdTable[Index + 1]) {
       DEBUG ((DEBUG_INFO, "PcdPei - SkuId is found in SkuId table.\n"));
       break;
